ver. 1.0.2-deprecated-fix

+++ kottsov.daniil/T3/polygon.h
@@ -52,7 +52,7 @@ namespace poly
   std::istream& operator>>(std::istream& in, DefInput&& def);
   struct NumberInput
   {
-    int& ref;
+    size_t& ref;
   };
   std::istream& operator>>(std::istream& in, NumberInput&& num);
 };
